api: start implementing Ptex Reader APIs
Add bindings for ptex::reader::Cache so that we can use
cached ptex readers from Rust. The reader is not working
yet but these are the first steps towards reading.

+++ bind/c-ptexture.cpp
@@ -266,7 +266,7 @@ struct PtexErrorHandler {
 struct PtexCache {
     using BoundType = Ptex::PtexCache;
 
-    static auto create(int maxFiles, unsigned long maxMem, bool premultiply,
+    static auto create(int maxFiles, size_t maxMem, bool premultiply,
                        Ptex::PtexInputHandler* inputHandler,
                        Ptex::PtexErrorHandler* errorHandler)
         -> Ptex::PtexCache*;
